Update HYGOV.mo
Changed default water head value to be a user-settable parameter. The value is also set now correctly in the constant source within the text layer to allow for initialization for a range of values of h. The value must be larger than zero to avoid division by zero error, could be handled better by the model, but for now I have documented the requirement in the parameter information.

Co-Authored-By: Marcelo de Castro <decasm3@rpi.edu>