Further updates to the IEEE Microgrid model
The "clean" version of the model now consists of components that are PSSE only (there are no multi-domain components here). For now I will continue this way. The system, however, does not contain renewable components nor motor and VSD. For now, we have a diesel generator, cap banks and load. The model also has records now that will facilitate our simulations in the near future.