Adding "--daeMode", "DAEsolver=true" and related changes
As per #305, I have finally added the openmodelica flag for DAE mode. I've also added (when missing) the DAEsolver=true flag for Dymola, and I've removed RK2 fixed solver settings for examples that would create an issue when attempting to run the model in Dymola using DAE mode (as those settings are contradicting).
Modified the documentation in Examples.DAE to cite the paper, which was added in the Reference sub-package with DOI and URL pointing to the paper.